Eliminate all acceptance testing
Considering the enormous amount of trouble we have testing mapbox - and the difficulty with mocking mapbox-gl's background assets for 100% offline testing - my thinking right now is that we skip acceptance testing entirely, and focus on unit and integration.

My thinking here is that it's easy enough to inject mapbox-gl as a dependency in our integration tests if we split up our components enough (read: no one component is doing too much). This depends a huge amount on wrapping critical app functional in components, especially when that component functionality is being passed to a component block.

The reason for that has to do with the dependency injection pattern - you can inject the mapbox-gl object into those integration tests, but you can run into problems if that component is passing around that object way too much.