417: Forces user to add a polygon label
When a user has just created a polygon and the label form is being
displayed, we want to force the user to select a polygon. This is
achieved by disabling the user to switch modes, either by selecting a
new annotation mode or selecting another feature/blank map area. A
top-level property, missingLabel, added to the selected feature is used
to conditionally display the label form as red when the user is blocked
from switching modes to cue them as to why they're stuck

Adds polygon label validation to isReadyToProceed check

Adds label enforcement to click action for polygon draw button